+++ src/main/java/eins/dao/CompanyUserDao.java
@@ -3,6 +3,6 @@ package eins.dao;
 import eins.entity.CompanyUser;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CompanyUserDao extends JpaRepository<CompanyUser,Integer> {
+public interface CompanyUserDao extends JpaRepository<CompanyUser, Long> {
 
 }
